requirements.txt
# =================================================
# Core Python & Utilities
# =================================================
numpy==1.26.4
tqdm==4.66.2
Pillow==10.2.0
scipy==1.12.0

# =================================================
# Computer Vision
# =================================================
opencv-python==4.9.0.80

# =================================================
# Video & Audio Processing
# =================================================
moviepy==1.0.3
torchaudio==2.2.1
soundfile==0.12.1

# =================================================
# Deep Learning (PyTorch - CPU Safe)
# =================================================
torch==2.2.1
torchvision==0.17.1

# =================================================
# Face Detection
# =================================================
facenet-pytorch==2.5.3

# =================================================
# Backend (FastAPI)
# =================================================
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pydantic==2.6.4

# =================================================
# NLP / Fraud Detection Support (SAFE)
# =================================================
pandas==2.2.1
scikit-learn==1.4.1.post1
joblib==1.3.2
nltk==3.8.1
langdetect==1.0.9

# =================================================
# Optional / Stability
# =================================================
regex==2023.12.25

SpeechRecognition==3.10.1

